+++ src/scenario/event/action_handler.cpp
@@ -109,9 +109,9 @@ void scenario_action_type_delete(scenario_action_t *action)
 
 void scenario_action_type_save_state(buffer *buf, const scenario_action_t *action, int link_type, int32_t link_id)
 {
-    buffer_write_i16(buf, link_type);
+    buffer_write_i16(buf, static_cast<int16_t>(link_type));
     buffer_write_i32(buf, link_id);
-    buffer_write_i16(buf, action->type);
+    buffer_write_i16(buf, static_cast<int16_t>(action->type));
     buffer_write_i32(buf, action->parameter1);
     buffer_write_i32(buf, action->parameter2);
     buffer_write_i32(buf, action->parameter3);
